requirements.txt
torch>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
transformers>=4.30.0
accelerate>=0.20.0
llama-cpp-python>=0.2.11
huggingface-hub>=0.19.0
requests>=2.31.0
PyPDF2>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2 